feat(ai-plugin-toolkit): new package; move plugin API out of @tanstack/ai (remove ./plugin subpath)
