+++ packages/ai-plugin-toolkit/vite.config.ts
@@ -29,14 +29,7 @@ const config = defineConfig({
 export default mergeConfig(
   config,
   tanstackViteConfig({
-    entry: [
-      './src/index.ts',
-      './src/client.ts',
-      './src/activities/index.ts',
-      './src/middlewares/index.ts',
-      './src/middlewares/otel.ts',
-      './src/adapter-internals.ts',
-    ],
+    entry: ['./src/index.ts'],
     srcDir: './src',
     cjs: false,
   }),
