+++ packages/ai/package.json
@@ -25,10 +25,6 @@
       "types": "./dist/esm/index.d.ts",
       "import": "./dist/esm/index.js"
     },
-    "./plugin": {
-      "types": "./dist/esm/plugin.d.ts",
-      "import": "./dist/esm/plugin.js"
-    },
     "./client": {
       "types": "./dist/esm/client.d.ts",
       "import": "./dist/esm/client.js"
